test: api 컨트롤러 테스트 추가
